% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_construct.R
\name{cohort_construct}
\alias{cohort_construct}
\title{Cohort constructor function (internal)}
\usage{
cohort_construct(start_year, end_year)
}
\arguments{
\item{start_year}{year of simulation start.}

\item{end_year}{year of simulation end}
}
\description{
Cohort constructor function (internal)
}
