% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_extract.R
\name{count_extract}
\alias{count_extract}
\title{count extractor function (internal)}
\usage{
count_extract(x, strategy = c("standard", "np1"))
}
\arguments{
\item{x}{list of run_model objects produced by model_runs function}

\item{strategy}{whether the standard or alternative strategy was used}
}
\description{
count extractor function (internal)
}
