% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_annotate.R
\name{data_annotate}
\alias{data_annotate}
\title{output data packaging function (internal)}
\usage{
data_annotate(x, start_year, end_year, pop_growth, init_size)
}
\arguments{
\item{x}{a list of list of count outcomes from the cout_extract function}

\item{start_year}{year of simulation start.}

\item{end_year}{year of simulation end}
}
\description{
output data packaging function (internal)
}
