% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_construct.R
\name{.cohort_construct}
\alias{.cohort_construct}
\title{Cohort constructor function (internal)}
\usage{
.cohort_construct(start_year, end_year, young_age, old_age)
}
\arguments{
\item{start_year}{year of simulation start.}

\item{end_year}{year of simulation end}

\item{young_age}{lowest starting age in simulation in years}

\item{old_age}{highest starting age in simulation in years}
}
\description{
Cohort constructor function (internal)
}
