% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_define.R
\name{.params_define}
\alias{.params_define}
\title{Parameters define function (internal)}
\usage{
.params_define(
  start_year,
  end_year,
  young_age,
  old_age,
  country,
  mar_prob,
  impact,
  sex = 0
)
}
\arguments{
\item{country}{three letter abbreviation of country WHO standard}

\item{mar_prob}{a vector of marriage probabilities (equal in length to runtime)}

\item{impact}{a number}

\item{age_init}{a numbers.}

\item{runtime}{a number (to represent the number of annual cycles of model)}
}
\description{
Parameters define function (internal)
}
