% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_define.R
\name{.strat_define}
\alias{.strat_define}
\title{strat define function (internal)}
\usage{
.strat_define(
  start_year,
  end_year,
  young_age,
  old_age,
  country,
  mar_prob,
  impact = 1.1,
  sex = 0,
  state_names = c("Unmarried", "Married", "Dead")
)
}
\arguments{
\item{country}{three letter abbreviation of country WHO standard}

\item{mar_prob}{a vector of marriage probabilities (equal in length to runtime)}

\item{impact}{a number}

\item{state_names}{a vector of state names}
}
\description{
strat define function (internal)
}
