% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_haz_define.R
\name{mar_haz_define}
\alias{mar_haz_define}
\title{marriage hazard construct function (internal)}
\usage{
mar_haz_define(start_year, end_year, mar_prob)
}
\arguments{
\item{start_year}{year of simulation start.}

\item{end_year}{year of simulation end}

\item{mar_prob}{list of data.frames of marriage probabilities by year (Starting at zero) thru maximum age
one for each cohort}
}
\description{
marriage hazard construct function (internal)
}
