% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runs.R
\name{model_runs}
\alias{model_runs}
\title{model runs function (internal)}
\usage{
model_runs(
  start_year,
  end_year,
  country,
  mar_prob,
  impact,
  sex = 0,
  state_names = c("Unmarried", "Married", "Dead"),
  pop_growth,
  init_size
)
}
\arguments{
\item{start_year}{year of simulation start.}

\item{end_year}{year of simulation end}

\item{country}{three letter abbreviation of country WHO standard}

\item{mar_prob}{a vector of marriage probabilities (equal in length to ncycles)}

\item{impact}{a number}

\item{sex}{sex for life table queries (0 = female) the default otherwise male}

\item{state_names}{a vector of state names}

\item{pop_growth}{population growth rate as a proportion}

\item{init_size}{population size for initial cohort}
}
\description{
model runs function (internal)
}
