% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_model.R
\name{summ_model}
\alias{summ_model}
\title{summarize adn weight data function (internal)}
\usage{
summ_model(x, type = c("cum_inc", "prevalence"))
}
\arguments{
\item{x}{a list of list of count outcomes from the cout_extract function}

\item{analytic_window}{a vector containing the start and end year to count marriages}
}
\description{
summarize adn weight data function (internal)
}
